+++ src/IpAddressVector.cpp
@@ -265,10 +265,10 @@ IpAddressVector IpAddressVector::createNetmask(IntegerVector in_pfx, LogicalVect
     if (in_v6[i] == NA_LOGICAL || in_pfx[i] == NA_INTEGER) {
       is_na[i] = true;
     } else if (in_v6[i]) {
-      address_v6[i] = get_netmask<asio::ip::address_v6>(in_pfx[i]);
+      address_v6[i] = prefix_to_netmask<asio::ip::address_v6>(in_pfx[i]);
       is_ipv6[i] = true;
     } else {
-      address_v4[i] = get_netmask<asio::ip::address_v4>(in_pfx[i]);
+      address_v4[i] = prefix_to_netmask<asio::ip::address_v4>(in_pfx[i]);
     }
   }
 
@@ -288,10 +288,10 @@ IpAddressVector IpAddressVector::createHostmask(IntegerVector in_pfx, LogicalVec
     if (in_v6[i] == NA_LOGICAL || in_pfx[i] == NA_INTEGER) {
       is_na[i] = true;
     } else if (in_v6[i]) {
-      address_v6[i] = get_hostmask<asio::ip::address_v6>(in_pfx[i]);
+      address_v6[i] = prefix_to_hostmask<asio::ip::address_v6>(in_pfx[i]);
       is_ipv6[i] = true;
     } else {
-      address_v4[i] = get_hostmask<asio::ip::address_v4>(in_pfx[i]);
+      address_v4[i] = prefix_to_hostmask<asio::ip::address_v4>(in_pfx[i]);
     }
   }
 
@@ -782,6 +782,43 @@ LogicalVector IpAddressVector::isWithinAny(const IpNetworkVector &network) const
         } else if (address_in_network(address_v4[i_addr], network.network_v4[i_netw])) {
           output[i_addr] = true;
         }
+      }
+    }
+  }
+
+  return output;
+}
+
+IntegerVector IpAddressVector::prefixFromMask() const {
+  std::size_t vsize = is_na.size();
+
+  // initialize vectors
+  IntegerVector output(vsize);
+
+  for (std::size_t i=0; i<vsize; ++i) {
+    if (is_na[i]) {
+      output[i] = NA_INTEGER;
+    } else if (is_ipv6[i]) {
+      int prefix = netmask_to_prefix(address_v6[i]);
+      if (prefix == -1) {
+        prefix = hostmask_to_prefix(address_v6[i]);
+      }
+      if (prefix == -1) {
+        warnInvalidInput(i, address_v6[i].to_string(), "invalid netmask/hostmask");
+        output[i] = NA_INTEGER;
+      } else {
+        output[i] = prefix;
+      }
+    } else {
+      int prefix = netmask_to_prefix(address_v4[i]);
+      if (prefix == -1) {
+        prefix = hostmask_to_prefix(address_v4[i]);
+      }
+      if (prefix == -1) {
+        warnInvalidInput(i, address_v4[i].to_string(), "invalid netmask/hostmask");
+        output[i] = NA_INTEGER;
+      } else {
+        output[i] = prefix;
       }
     }
   }